extract parseKafkaBatch unmodified before changes
